+++ packages/react-ui/.storybook/decorators/Theme/ThemeDecorator.tsx
@@ -3,8 +3,10 @@ import { Decorator } from '@storybook/react';
 
 import { LIGHT_THEME } from '../../../lib/theming/themes/LightTheme';
 import { DARK_THEME } from '../../../lib/theming/themes/DarkTheme';
-import { LIGHT_THEME_2022_0 } from '../../../lib/theming/themes/LightTheme2022_0';
-import { DARK_THEME_2022_0 } from '../../../lib/theming/themes/DarkTheme2022_0';
+import { LIGHT_THEME_5_0 } from '../../../lib/theming/themes/LightTheme5_0';
+import { DARK_THEME_5_0 } from '../../../lib/theming/themes/DarkTheme5_0';
+import { LIGHT_THEME_5_1 } from '../../../lib/theming/themes/LightTheme5_1';
+import { DARK_THEME_5_1 } from '../../../lib/theming/themes/DarkTheme5_1';
 import { ThemeContext } from '../../../lib/theming/ThemeContext';
 import { ThemeFactory } from '../../../lib/theming/ThemeFactory';
 import { isDarkTheme } from '../../../lib/theming/ThemeHelpers';
@@ -12,8 +14,10 @@ import { isDarkTheme } from '../../../lib/theming/ThemeHelpers';
 export const themes = {
   LIGHT_THEME,
   DARK_THEME,
-  LIGHT_THEME_2022_0,
-  DARK_THEME_2022_0,
+  LIGHT_THEME_5_0,
+  DARK_THEME_5_0,
+  LIGHT_THEME_5_1,
+  DARK_THEME_5_1,
 };
 
 export const ThemeDecorator: Decorator = (Story, context) => {
