+++ packages/react-ui/index.ts
@@ -51,8 +51,10 @@ export * from './lib/featureFlagsContext';
 export * from './lib/locale';
 export * from './lib/theming/ThemeContext';
 export * from './lib/theming/ThemeFactory';
-export * from './lib/theming/themes/LightTheme2022_0';
-export * from './lib/theming/themes/DarkTheme2022_0';
+export * from './lib/theming/themes/LightTheme5_0';
+export * from './lib/theming/themes/DarkTheme5_0';
+export * from './lib/theming/themes/LightTheme5_1';
+export * from './lib/theming/themes/DarkTheme5_1';
 export * from './lib/theming/themes/LightTheme';
 export * from './lib/theming/themes/DarkTheme';
 export * from './lib/types/props';
