+++ packages/react-ui/internal/themes/BasicTheme.ts
@@ -1,7 +1,7 @@
+import { createThemeFromClass } from '../../lib/theming/ThemeHelpers';
 import * as ColorFunctions from '../../lib/styles/ColorFunctions';
-import { exposeGetters } from '../../lib/theming/ThemeHelpers';
 
-export class BasicLightThemeInternal {
+export class BasicThemeClass {
   //#region Common variables
   public static labGrotesqueBaselineCompensation = '1';
   public static brand = '#2291ff';
@@ -2484,4 +2484,4 @@ export class BasicLightThemeInternal {
   //#endregion
 }
 
-export const BasicLightTheme = exposeGetters(BasicLightThemeInternal);
+export const BasicTheme = createThemeFromClass(BasicThemeClass);
