+++ packages/react-ui/internal/themes/LightTheme5_0.ts
@@ -1,0 +1,8 @@
+import { createThemeFromClass, markThemeVersion } from '../../lib/theming/ThemeHelpers';
+
+import { BasicTheme, BasicThemeClass } from './BasicTheme';
+
+export const LightTheme5_0 = createThemeFromClass(class LightTheme5_0 extends (class {} as typeof BasicThemeClass) {}, {
+  prototypeTheme: BasicTheme,
+  themeMarkers: [markThemeVersion(5.0)],
+});
