+++ packages/react-ui/lib/theming/ThemeHelpers.ts
@@ -1,3 +1,4 @@
+import { BasicThemeClass } from '../../internal/themes/BasicTheme';
 import { isNonNullable } from '../utils';
 
 import { Theme, ThemeIn } from './Theme';
@@ -25,6 +26,10 @@ export const REACT_UI_THEME_MARKERS = {
   theme2022: {
     key: '__IS_REACT_UI_THEME_2022__',
     value: true,
+  },
+  themeVersion: {
+    key: '__REACT_UI_THEME_VERSION__',
+    value: 0,
   },
 };
 
@@ -58,9 +63,25 @@ export const markAsTheme2022: Marker = (theme) => {
   });
 };
 
+export const markThemeVersion: (version: number) => Marker = (version) => (theme) => {
+  return Object.create(theme, {
+    [REACT_UI_THEME_MARKERS.themeVersion.key]: {
+      value: version,
+      writable: false,
+      enumerable: false,
+      configurable: false,
+    },
+  });
+};
+
 export const isTheme2022 = (theme: Theme | ThemeIn): boolean => {
   // @ts-expect-error: internal value.
   return theme[REACT_UI_THEME_MARKERS.theme2022.key] === REACT_UI_THEME_MARKERS.theme2022.value;
+};
+
+export const isThemeVersionGreaterOrEqual = (theme: Theme | ThemeIn, version: number): boolean => {
+  // @ts-expect-error: internal value.
+  return theme[REACT_UI_THEME_MARKERS.themeVersion.key] >= version;
 };
 
 export function findPropertyDescriptor(theme: Theme, propName: string) {
@@ -76,9 +97,24 @@ export function findPropertyDescriptor(theme: Theme, propName: string) {
 }
 
 export function applyMarkers(theme: Readonly<Theme>, markers: Markers) {
-  let markedTheme: Readonly<Theme> = theme;
-  markers.forEach((marker) => {
-    markedTheme = marker(theme);
-  });
-  return markedTheme;
+  return markers.reduce((markedTheme, marker) => {
+    return marker(markedTheme);
+  }, Object.create(theme)) as typeof theme;
+}
+
+export function createThemeFromClass(
+  themeObject: typeof BasicThemeClass,
+  options?: {
+    prototypeTheme?: Theme;
+    themeMarkers?: Markers;
+  },
+): Theme {
+  const theme = exposeGetters(themeObject);
+  const { prototypeTheme, themeMarkers = [] } = options || {};
+
+  if (prototypeTheme) {
+    Object.setPrototypeOf(theme, prototypeTheme);
+  }
+
+  return applyMarkers(theme as Theme, themeMarkers);
 }
