+++ packages/react-ui/lib/theming/__tests__/Theming-test.tsx
@@ -2,11 +2,12 @@ import { render } from '@testing-library/react';
 import React from 'react';
 
 import { ThemeContext } from '../ThemeContext';
-import { applyMarkers, exposeGetters, Marker, REACT_UI_THEME_MARKERS } from '../ThemeHelpers';
+import { applyMarkers, createThemeFromClass, Marker, REACT_UI_THEME_MARKERS } from '../ThemeHelpers';
 import { ThemeFactory } from '../ThemeFactory';
 import { Theme } from '../Theme';
-import { BasicLightThemeInternal, BasicLightTheme } from '../../../internal/themes/BasicLightTheme';
+import { LIGHT_THEME } from '../../../lib/theming/themes/LightTheme';
 import { AnyObject } from '../../utils';
+import { BasicTheme, BasicThemeClass } from '../../../internal/themes/BasicTheme';
 
 const TEST_MARKERS = {
   test: {
@@ -34,9 +35,12 @@ const getConsumedTheme = () => {
 
 // test theme
 const myTheme = { brand: 'custom', bgDefault: 'custom' } as const;
-const TestTheme = Object.setPrototypeOf(
-  exposeGetters({ bgDefault: 'default', bgSecondary: 'default' }),
-  BasicLightTheme,
+const TestTheme = createThemeFromClass(
+  class extends (class {} as typeof BasicThemeClass) {
+    public static bgDefault = 'default';
+    public static bgSecondary = 'default';
+  },
+  { prototypeTheme: BasicTheme },
 );
 
 // test marker
@@ -63,7 +67,7 @@ describe('Theming', () => {
         const theme = ThemeFactory.create(myTheme);
 
         expect(theme.brand).toEqual(myTheme.brand);
-        expect(theme.black).toEqual(BasicLightTheme.black);
+        expect(theme.black).toEqual(LIGHT_THEME.black);
       });
       test('with args [theme, baseTheme]', () => {
         const theme = ThemeFactory.create(myTheme, TestTheme);
@@ -93,7 +97,7 @@ describe('Theming', () => {
     });
     test('getKeys()', () => {
       const keys_1 = ThemeFactory.getKeys(TestTheme);
-      const keys_2 = ThemeFactory.getKeys(BasicLightThemeInternal);
+      const keys_2 = ThemeFactory.getKeys(LIGHT_THEME);
 
       expect(keys_1).toEqual(keys_2);
     });
