+++ packages/react-ui/lib/theming/themes/DarkTheme.ts
@@ -1,5 +1,1 @@
-import { ThemeFactory } from '../ThemeFactory';
-import { BasicDarkTheme } from '../../../internal/themes/BasicDarkTheme';
-import { applyMarkers, markAsDarkTheme, markAsTheme2022 } from '../ThemeHelpers';
-
-export const DARK_THEME = applyMarkers(ThemeFactory.create({}, BasicDarkTheme), [markAsTheme2022, markAsDarkTheme]);
+export { DarkTheme5_0 as DARK_THEME } from '../../../internal/themes/DarkTheme5_0';
