--- packages/react-ui/lib/theming/themes/DarkTheme2022_0.ts
@@ -1,5 +1,0 @@
-import { ThemeFactory } from '../ThemeFactory';
-
-import { DARK_THEME } from './DarkTheme';
-
-export const DARK_THEME_2022_0 = ThemeFactory.create({}, DARK_THEME);
