+++ packages/react-ui/lib/theming/themes/DarkTheme5_0.ts
@@ -1,0 +1,1 @@
+export { DarkTheme5_0 as DARK_THEME_5_0 } from '../../../internal/themes/DarkTheme5_0';
