+++ packages/react-ui/lib/theming/themes/DarkTheme5_1.ts
@@ -1,0 +1,1 @@
+export { DarkTheme5_1 as DARK_THEME_5_1 } from '../../../internal/themes/DarkTheme5_1';
