--- packages/react-ui/lib/theming/themes/LightTheme2022_0.ts
@@ -1,5 +1,0 @@
-import { ThemeFactory } from '../ThemeFactory';
-
-import { LIGHT_THEME } from './LightTheme';
-
-export const LIGHT_THEME_2022_0 = ThemeFactory.create({}, LIGHT_THEME);
