+++ packages/react-ui/lib/theming/themes/LightTheme5_0.ts
@@ -1,0 +1,1 @@
+export { LightTheme5_0 as LIGHT_THEME_5_0 } from '../../../internal/themes/LightTheme5_0';
