+++ packages/react-ui/lib/theming/themes/LightTheme5_1.ts
@@ -1,0 +1,1 @@
+export { LightTheme5_1 as LIGHT_THEME_5_1 } from '../../../internal/themes/LightTheme5_1';
